+++ src/packages/piral-cli/src/common/SharedDependenciesAsset.ts
