+++ src/packages/piral-cli/src/common/VirtualAsset.test.ts
