+++ src/packages/piral-cli/src/common/VirtualAsset.ts
