+++ src/packages/piral-cli/src/common/VirtualPackager.test.ts
