+++ src/packages/piral-cli/src/common/emoji.ts
