+++ src/packages/piral-cli/src/common/hash.test.ts
