+++ src/packages/piral-cli/src/common/hash.ts
