+++ src/packages/piral-cli/src/common/info.ts
