+++ src/packages/piral-cli/src/common/inspect.ts
