+++ src/packages/piral-cli/src/common/npm.test.ts
