+++ src/packages/piral-cli/src/common/pack.ts
