+++ src/packages/piral-cli/src/common/scaffoldPilet.ts
