+++ src/packages/piral-cli/src/common/scaffoldPiral.ts
